what is changed: you can now input -go-method 1 to use the new implementation (0 for the old one) this method assumes that the molecules in your pdb file are all their own chain.
martinize2
you should also give a martini3 ff in argument -go-ff so that the correct LJ sigma and epsilon can be read for each non-bond params cancellation
you can set cutoffs and epsilon values for both inter and intramolecular bonds (with -go-low-intra/inter and -go-up-intra/inter for cutoffs, go-eps-intra/inter for epsilon values)

NOTE: the new implementation is not yet tested for when you give the contact map with -go ‘contact map path’
With the new implementation, the system is not merged, so that separate itp files can be written for unique molecules (e.g. heterodimers). The merged system is generated, because it is needed for the generation of go contacts.
However the resids of the contacts that are gotten from the merged system are not the same as the system where all molecules are separate. So there is a mapping based on atomid. The mapping dict in martinize2 is called map_after_martinize. And is added as an attribute in the nodes of the merged system as [‘resid_before_merge’]. This is added to the merged system, because the go contacts are generated with the merged system (in go_structure_bias.py)

The last that is changed in the martinize2 is the amount of arguments given to the GoPipeline.run_system(), namely the extra arguments that were mentioned (-go-ff, -go-low-intra/inter, etc.)