go_pipeline.py Both methods have their own prepare_run, in method 1 (the new one) the unique molecules in the system are identified, so that all other molecules that are the same, have the same mol.meta[‘moltype’] (= f"{moltype}_{reference_chain_id}"). As you can see the meta[‘moltype’] of each homologous molecule is set to the reference_chain_id, which is the name of the first chain in this set of homologous molecules. This is important because then only 1 itp file is written for each unique molecule.
There are some dictionaries made for mapping in the 2 systems that are used (merged and non-merged). these mappings are put into system.go_params
The first is chain_to_mol_id, which gives a mapping of the chain (‘A’, ‘B’, ‘C’, etc.) to the mol id in the system (0, 1, 2, etc.)
The second is unique_mols_map, which gives a mapping for each chain, to all the chains that are homologous/the same (e. g. {'A': ['B', 'C', 'D'], 'B': ['A', 'C', 'D'], 'C': ['A', 'B', 'D'], 'D': ['A', 'B', 'C'], 'E': []})
The third is reference_chain_map, which maps the chain to the ‘mother chain’, which is the reference_chain_id that was mentioned earlier. (e.g. {'A': 'A', 'B': 'A', 'C': 'A', 'D': 'A', 'E': 'E'})

Then as usual the system is put through the 3 processes. SetMoleculeMeta, VirtualSiteCreator and ComputeStructuralGoBias, which point to the set_molecule_meta.py, go_vs_includes.py and go_structure_bias.py scripts.
in the set_molecule_meta.py nothing was changed, it just correctly sets the meta of each molecule in the system.